Add cmake module DoxygenGitVersion
`DoxygenGitVersion.cmake` allows to add a hook to `doxygen_add_docs()`
to automatically update the version in the documentation
(`PROJECT_NUMBER`) according to the information in `Git`.
The current version relies on internal behavior of the Doxygen module
in CMake (name and location of the Doxyfile).

Example snippet:

        include(DoxygenGitVersion.cmake)
        doxygen_add_docs(doxygen_doc)
        add_doxygen_version_hook(doxygen_doc)

(This module has to be added as sub-directory before)